+++ src/common/helpers/inventory.ts
@@ -1,4 +1,3 @@
-import { Action } from 'org.bukkit.event.block';
 import { Player } from 'org.bukkit.entity';
 import {
   ItemStack,
@@ -6,32 +5,6 @@ import {
   PlayerInventory,
 } from 'org.bukkit.inventory';
 import { Material } from 'org.bukkit';
-
-/**
- * @param chance Chance of success. 0-1
- * @returns True of false based on the chance.
- */
-export function chanceOf(chance: number) {
-  return Math.random() < chance;
-}
-
-/**
- * @param action Action of the click event.
- * @returns True of false.
- */
-export function isLeftClick(action: Action) {
-  return action === Action.LEFT_CLICK_AIR || action === Action.LEFT_CLICK_BLOCK;
-}
-
-/**
- * @param action Action of the click event.
- * @returns True of false.
- */
-export function isRightClick(action: Action) {
-  return (
-    action === Action.RIGHT_CLICK_AIR || action === Action.RIGHT_CLICK_BLOCK
-  );
-}
 
 /**
  * Gives item to the player or drops it on the ground, if the inventory is full.
